requirements.txt
astapi==0.115.0
uvicorn==0.29.0
pydantic==2.9.0
python-jose
passlib
python-multipart
sqlalchemy
psycopg2-binary
pytest
httpx